requirements.txt
django==3.2.5
Pillow
wikipedia
gtts
pydub
backgroundremover
numpy<=1.20
dj-database-url
gunicorn